// Frontend/src/app/core/models/product/product-sub-category.ts
export enum ProductSubCategory {
  // Subcategorías de Technology
  PC = 'PC',
  Console = 'Console',
  Smartphone = 'Smartphone',

  // Subcategorías de Clothing
  Men = 'Men',
  Women = 'Women',
  Kids = 'Kids',

  // Subcategorías de Home
  Furniture = 'Furniture',
  Kitchen = 'Kitchen',
  Decor = 'Decor',
}